refactor: 서버 유틸 함수들 분리
